most tests pass with relative imports, just need to check test_unselect since there might be some cyclic import problems where user imports create condition which imports user...
