+++ backend/algorithms/objects/categories.py
@@ -125,33 +125,3 @@ class LevelCourseCategory(Category):
 
         return total_wam / total_uoc
 
-
-def create_category(tokens):
-    '''Given a list of tokens starting from after the connector keyword, create
-    and return the category object matching the category, as well as the current index
-    of the token list.'''
-
-    # At most we will only parse 1 or 2 tokens so no need for an iterator
-    # NOTE: There will always be at least 2 tokens due to a closing ")" bracket
-    # so it is safe to check tokens[1]
-    if re.match(r'^[A-Z]{4}$', tokens[0], flags=re.IGNORECASE):
-        # Course type
-        return CourseCategory(tokens[0]), 0
-    elif re.match(r'^L[0-9]$', tokens[0], flags=re.IGNORECASE):
-        # Level category. Get the level, then determine next token if there is one
-        level = int(re.match(r'^L([0-9])$', tokens[0],
-                    flags=re.IGNORECASE).group(1))
-
-        if re.match(r'^[A-Z]{4}$', tokens[1], flags=re.IGNORECASE):
-            # Level Course Category. e.g. L2 MATH
-            course_code = re.match(r'^([A-Z]{4})$', tokens[1], flags=re.IGNORECASE).group(1)
-
-            return LevelCourseCategory(level, course_code), 1
-        else:
-            # There are no tokens after this. Simple level category
-            return LevelCategory(level), 0
-
-    # TODO: Levels (e.g. L2 MATH, SPECIALISATIONS, PROGRAM)
-
-    # Did not match any category. Return None and assume only 1 token was consumed
-    return None, 0
